do some crazy hacks in order to get the grading script for PA5 to work
Lots of paths are hardcoded in the provided scripts and binaries and
as a result the grading script for PA5 doesn't work out of the box.
The changes introduced in this patch fix the problem but also *force*
the project to be placed under /usr/class as intented by the course
creators.